+++ buildSrc/Builder.js
@@ -8,7 +8,7 @@ import nodeResolve from "@rollup/plugin-node-resolve"
 import hmr from "nollup/lib/plugin-hmr.js"
 import os from "os"
 import {bundleDependencyCheckPlugin} from "./RollupConfig.js"
-import {babelDesktopPlugins} from "./RollupConfig.js"
+import {nativeDepWorkaroundPlugin, pluginNativeLoader, babelDesktopPlugins} from "./RollupPlugins.js"
 
 const __dirname = dirname(fileURLToPath(import.meta.url))
 const root = path.dirname(__dirname)
@@ -203,53 +203,4 @@ async function buildAndStartDesktop(log, version) {
 
 	log("Bundled desktop client")
 	return [nodeBundleWrapper]
-}
-
-export function nativeDepWorkaroundPlugin() {
-	return {
-		name: "native-dep-workaround",
-		resolveId(id) {
-			// It's included in the build by electron builder, consider it external
-			if (id === "electron") {
-				return false
-			}
-			// There are issues with packaging it so we include it as unprocessed cjs. It doesn't tree-shake well anyway.
-			if (id === "electron-updater") {
-				return false
-			}
-			// We currently have an import in Rsa.js which we don't want in Desktop as it pulls the whole worker with it
-			if (id.endsWith("RsaApp")) {
-				return false
-			}
-		}
-	}
-}
-
-/**
- * This plugin loads native node module (.node extension).
- * This is not general enough yet, it only works in commonjs and it doesn't use ROLLUP_ASSET_URL.
- * This will also not work with async imports.
- *
- * Important! Make sure that requireReturnsDefault for commonjs plugin is set to `true` or `"preferred"` if .node module is part of
- * commonjs code.
- */
-export function pluginNativeLoader() {
-	return {
-		name: "native-loader",
-		async load(id) {
-			if (id.endsWith(".node")) {
-				const name = path.basename(id)
-				const content = await fs.promises.readFile(id)
-				this.emitFile({
-					type: 'asset',
-					name,
-					fileName: name,
-					source: content,
-				})
-				return `
-				const nativeModule = require('./${name}')
-				export default nativeModule`
-			}
-		},
-	}
 }
